Extract ChaosMetrics out from network.h
