+++ flow/flat_buffers.cpp
@@ -27,6 +27,7 @@
 
 #include <algorithm>
 #include <iomanip>
+#include <iostream>
 #include <unordered_set>
 #include <variant>
