+++ flow/include/flow/network.h
@@ -37,12 +37,6 @@
 
 class Void;
 
-
-template <>
-struct Traceable<IPAddress> : std::true_type {
-	static std::string toString(const IPAddress& value) { return value.toString(); }
-};
-
 FDB_DECLARE_BOOLEAN_PARAM(NetworkAddressFromHostname);
 
 struct NetworkAddress {
@@ -590,120 +584,4 @@ public:
 	// Returns the interface that should be used to make and accept socket connections
 };
 
-// Chaos Metrics - We periodically log chaosMetrics to make sure that chaos events are happening
-// Only includes DiskDelays which encapsulates all type delays and BitFlips for now
-// Expand as per need
-struct ChaosMetrics {
-
-	ChaosMetrics() { clear(); }
-
-	void clear() {
-		memset(this, 0, sizeof(ChaosMetrics));
-		startTime = g_network ? g_network->now() : 0;
-	}
-
-	unsigned int diskDelays;
-	unsigned int bitFlips;
-	double startTime;
-
-	void getFields(TraceEvent* e) {
-		std::pair<const char*, unsigned int> metrics[] = { { "DiskDelays", diskDelays }, { "BitFlips", bitFlips } };
-		if (e != nullptr) {
-			for (auto& m : metrics) {
-				char c = m.first[0];
-				if (c != 0) {
-					e->detail(m.first, m.second);
-				}
-			}
-		}
-	}
-};
-
-// This class supports injecting two type of disk failures
-// 1. Stalls: Every interval seconds, the disk will stall and no IO will complete for x seconds, where x is a randomly
-// chosen interval
-// 2. Slowdown: Random slowdown is injected to each disk operation for specified period of time
-struct DiskFailureInjector {
-	static DiskFailureInjector* injector() {
-		auto res = g_network->global(INetwork::enDiskFailureInjector);
-		if (!res) {
-			res = new DiskFailureInjector();
-			g_network->setGlobal(INetwork::enDiskFailureInjector, res);
-		}
-		return static_cast<DiskFailureInjector*>(res);
-	}
-
-	void setDiskFailure(double interval, double stallFor, double throttleFor) {
-		stallInterval = interval;
-		stallPeriod = stallFor;
-		stallUntil = std::max(stallUntil, g_network->now() + stallFor);
-		// random stall duration in ms (chosen once)
-		// TODO: make this delay configurable
-		stallDuration = 0.001 * deterministicRandom()->randomInt(1, 5);
-		throttlePeriod = throttleFor;
-		throttleUntil = std::max(throttleUntil, g_network->now() + throttleFor);
-		TraceEvent("SetDiskFailure")
-		    .detail("Now", g_network->now())
-		    .detail("StallInterval", interval)
-		    .detail("StallPeriod", stallFor)
-		    .detail("StallUntil", stallUntil)
-		    .detail("ThrottlePeriod", throttleFor)
-		    .detail("ThrottleUntil", throttleUntil);
-	}
-
-	double getStallDelay() {
-		// If we are in a stall period and a stallInterval was specified, determine the
-		// delay to be inserted
-		if (((stallUntil - g_network->now()) > 0.0) && stallInterval) {
-			auto timeElapsed = fmod(g_network->now(), stallInterval);
-			return std::max(0.0, stallDuration - timeElapsed);
-		}
-		return 0.0;
-	}
-
-	double getThrottleDelay() {
-		// If we are in the throttle period, insert a random delay (in ms)
-		// TODO: make this delay configurable
-		if ((throttleUntil - g_network->now()) > 0.0)
-			return (0.001 * deterministicRandom()->randomInt(1, 3));
-
-		return 0.0;
-	}
-
-	double getDiskDelay() { return getStallDelay() + getThrottleDelay(); }
-
-private: // members
-	double stallInterval = 0.0; // how often should the disk be stalled (0 meaning once, 10 meaning every 10 secs)
-	double stallPeriod; // Period of time disk stalls will be injected for
-	double stallUntil; // End of disk stall period
-	double stallDuration; // Duration of each stall
-	double throttlePeriod; // Period of time the disk will be slowed down for
-	double throttleUntil; // End of disk slowdown period
-
-private: // construction
-	DiskFailureInjector() = default;
-	DiskFailureInjector(DiskFailureInjector const&) = delete;
-};
-
-struct BitFlipper {
-	static BitFlipper* flipper() {
-		auto res = g_network->global(INetwork::enBitFlipper);
-		if (!res) {
-			res = new BitFlipper();
-			g_network->setGlobal(INetwork::enBitFlipper, res);
-		}
-		return static_cast<BitFlipper*>(res);
-	}
-
-	double getBitFlipPercentage() { return bitFlipPercentage; }
-
-	void setBitFlipPercentage(double percentage) { bitFlipPercentage = percentage; }
-
-private: // members
-	double bitFlipPercentage = 0.0;
-
-private: // construction
-	BitFlipper() = default;
-	BitFlipper(BitFlipper const&) = delete;
-};
 #endif
